ImgPlane: move inline funcs to header
Will be called by CASPER implementation.
